Added ErrorViewModel, Constants and CustomError view

+++ src/Web/IdeGames.Web/Controllers/NewsController.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using IdeGames.Services.Contracts;
 using IdeGames.Services.Models.Models.News;
+using IdeGames.Web.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -26,7 +27,12 @@ namespace IdeGames.Web.Controllers
         public IActionResult Delete(int id)
         {
             var news = this.Db.News.FirstOrDefault(x => x.Id == id);
-            
+
+            if (news == null)
+            {
+                return this.CustomError(Constants.NullModelError);
+            }
+
             this.Db.Remove(news);
 
             this.Db.SaveChanges();
@@ -49,7 +55,7 @@ namespace IdeGames.Web.Controllers
 
             this.Db.SaveChanges();
             
-            return this.RedirectToPage("");
+            return this.RedirectToAction("Create");
         }
     }
 }
